Design refactor
