--- exercise1/problema1.c
@@ -1,10 +1,0 @@
-#define Voltage 220
-#define current 10
-#define <stdio.h>
-
-main()
-{
-float Resistor;
-Resistor = Voltage/current;
-printf("R = %6.2f\n", Resistor);
-}
