--- exercise1/problema3.c
@@ -1,13 +1,0 @@
-#include <stdio.h>
-
-extern float resistor_fun(float);
-
-int main(){
-    float resistor, current;
-    printf("Enter the current value: ");
-    scanf("%f", &current);
-
-    resistor = resistor_fun(current);
-    printf("The required resistor should be ");
-    printf("%1.0f Ohmns.\n", resistor);
-}
