--- exercise1/problema4_2.h
@@ -1,12 +1,0 @@
-int the_biggest(int x, int y){
-    if(x > y)
-        return x;
-    else
-        return y;
-}
-int the_smallest(int x, int y){
-    if(x < y)
-        return x;
-    else
-        return y;
-}
