+++ exercises/exercise1/pre_problema.c
@@ -1,0 +1,16 @@
+# 1 "problema.c"
+# 1 "<built-in>" 1
+# 1 "<built-in>" 3
+# 362 "<built-in>" 3
+# 1 "<command line>" 1
+# 1 "<built-in>" 2
+# 1 "problema.c" 2
+
+
+
+main()
+{
+float Resistor;
+Resistor = 220/10;
+printf("R = %6.2f\n", Resistor);
+}
