+++ exercises/exercise1/problema3_b.c
@@ -1,0 +1,7 @@
+#define Voltage 5
+
+float resistor_fun(float i){
+    float b;
+    b = Voltage/i;
+    return b;
+}
