+++ exercises/exercise1/problema4.c
@@ -1,0 +1,8 @@
+#define BIGNUM 1000000
+#include <stdio.h>
+
+int main(void){
+    int a = BIGNUM;
+    printf("%d\n",a);
+    return 0;
+}
