+++ exercises/exercise1/problema4_2.c
@@ -1,0 +1,10 @@
+#include <stdio.h>
+#include "problema4_2.h"
+
+int main(void){
+    int a = 10;
+    int b = 4;
+    printf("The biggest is: %d\n", the_biggest(a,b));
+    printf("The smallest is: %d\n", the_smallest(a,b));
+    return 0;
+}
