+++ exercises/exercise1/problema4_2.h
@@ -1,0 +1,12 @@
+int the_biggest(int x, int y){
+    if(x > y)
+        return x;
+    else
+        return y;
+}
+int the_smallest(int x, int y){
+    if(x < y)
+        return x;
+    else
+        return y;
+}
