+++ exercises/exercise3/eje.txt
@@ -1,0 +1,18 @@
+AC3
+P1-22
+( * (+ 1 39) (- 53 45))
+(+ (/ 1020 39) (* 45 2))
+(+ 39 48 72 23 91)
+(/ (+ 39 48 72 23 91) 5.0)
+
+(log 1024 2)
+(log 100)
+(expt 2 3)
+(exp 2)
+
+(define pi 3.141592)
+
+(definir area de un circulo de radio)
+
+
+(* pi 2 2)
