+++ exercises/exercise4/mayor.scm
@@ -1,0 +1,16 @@
+(define (mayor a b)
+    (cond 
+    [( > a b) b ]
+    [else b]
+    )
+)
+
+(if (> 1 0) #t #f)
+
+(define (!= a b)
+    (not (= a b))
+)
+
+(and #t #f #t)
+
+(or #t #f)
